+++ examples/ruby-sample/tests/integration_test.rs
@@ -7,8 +7,7 @@
 // https://rust-lang.github.io/rust-clippy/stable/index.html
 #![warn(clippy::pedantic)]
 
-use libcnb_test::assert_contains;
-use libcnb_test::IntegrationTest;
+use libcnb_test::{assert_contains, assert_not_contains, TestConfig, TestRunner};
 use std::io;
 use std::io::{Read, Write};
 use std::net;
@@ -18,35 +17,40 @@ use std::time::Duration;
 #[test]
 #[ignore]
 fn basic() {
-    IntegrationTest::new("heroku/buildpacks:20", "test-fixtures/simple-ruby-app").run_test(
-        |context| {
-            assert_contains!(context.pack_stdout, "---> Ruby Buildpack");
-            assert_contains!(context.pack_stdout, "---> Installing bundler");
-            assert_contains!(context.pack_stdout, "---> Installing gems");
+    let config = TestConfig::new("heroku/buildpacks:20", "test-fixtures/simple-ruby-app");
 
-            context
-                .prepare_container()
-                .env("PORT", TEST_PORT.to_string())
-                .expose_port(TEST_PORT)
-                .start_with_default_process(|container| {
-                    std::thread::sleep(Duration::from_secs(1));
+    TestRunner::default().run_test(&config, |context| {
+        assert_contains!(context.pack_stdout, "---> Ruby Buildpack");
+        assert_contains!(context.pack_stdout, "---> Installing bundler");
+        assert_contains!(context.pack_stdout, "---> Installing gems");
 
-                    assert_eq!(
-                        call_test_fixture_service(
-                            container.address_for_port(TEST_PORT).unwrap(),
-                            "Hello World!"
-                        )
-                        .unwrap(),
-                        "!dlroW olleH"
-                    );
+        context
+            .prepare_container()
+            .env("PORT", TEST_PORT.to_string())
+            .expose_port(TEST_PORT)
+            .start_with_default_process(|container| {
+                std::thread::sleep(Duration::from_secs(1));
 
-                    assert_contains!(
-                        container.shell_exec("ruby --version").stdout,
-                        "ruby 2.7.0p0"
-                    );
-                });
-        },
-    );
+                assert_eq!(
+                    call_test_fixture_service(
+                        container.address_for_port(TEST_PORT).unwrap(),
+                        "Hello World!"
+                    )
+                    .unwrap(),
+                    "!dlroW olleH"
+                );
+
+                assert_contains!(
+                    container.shell_exec("ruby --version").stdout,
+                    "ruby 2.7.0p0"
+                );
+            });
+
+        context.run_test(&config, |context| {
+            assert_not_contains!(context.pack_stdout, "---> Installing bundler");
+            assert_not_contains!(context.pack_stdout, "---> Installing gems");
+        });
+    });
 }
 
 fn call_test_fixture_service<A>(a: A, payload: impl AsRef<str>) -> io::Result<String>